+++ src/main.cpp
@@ -369,6 +369,8 @@ int main(int argc, char* argv[]) {
       config.client_id = args.ayame_client_id;
       config.signaling_key = args.ayame_signaling_key;
       config.direction = args.ayame_direction;
+      config.video_codec_type = args.ayame_video_codec_type;
+      config.audio_codec_type = args.ayame_audio_codec_type;
 
       ayame_client =
           AyameClient::Create(ioc, rtc_manager.get(), std::move(config));