+++ src/momo_args.h
@@ -87,6 +87,9 @@ struct MomoArgs {
   std::string ayame_signaling_key = "";
   // sendrecv, sendonly, recvonly
   std::string ayame_direction = "sendrecv";
+  // 空文字の場合コーデックは WebRTC デフォルトを使用
+  std::string ayame_video_codec_type = "";
+  std::string ayame_audio_codec_type = "";
 
   bool disable_echo_cancellation = false;
   bool disable_auto_gain_control = false;